Problema "Terreno"  | "Terrain" problem
This was my second exercise in my Algorithms and Programming Logic course.

I needed to make a program that first reads the measurements of the width and length of a rectangular piece of land (to one decimal place), as well as the value of the square meter of the land (to two decimal places). Then, the program should show the value of the area of the land, as well as the value of the price of the land (both to two decimal places). And so I did it! <3